Introduce BaseShopItemView
