+++ nekoyume/Assets/_Scripts/UI/Shop/BaseShopItemView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4c5e644bbd1544428344a76428506e24
+timeCreated: 1644977811
